# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Torrent streaming server.",
        "title": "Swagger Torrserver API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "{version.Version}"
    },
    "basePath": "/",
    "paths": {
...
        "/ffp/status": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Reports whether the ffprobe binary is available. Features that need real\nmedia duration (e.g. saving playback position) are only usable when true.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "ffprobe availability",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ffprobeStatusResponse"
                        }
                    }
                }
            }
        },
        "/ffp/{hash}/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Gather informations using ffprobe.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "Gather informations using ffprobe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Torrent hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File index in torrent",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data returned from ffprobe"
                    }
                }
            }
        },
...
        "/gst/{hash}/probe": {
            "get": {
                "description": "Requires a `-gst` build. Uses gst-discoverer to inspect container and codec information.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GStreamer"
                ],
                "summary": "Probe torrent media tracks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Torrent infohash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "File index in torrent",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "File index (alias)",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "File index (alias)",
                        "name": "fileID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Media probe result",
                        "schema": {
                            "$ref": "#/definitions/api.gstreamerProbeInfoDoc"
                        }
                    },
                    "400": {
                        "description": "Invalid source",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Probe failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "504": {
                        "description": "Probe timed out",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.ffprobeStatusResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                }
            }
        },
...
        "api.gstreamerProbeInfoDoc": {
            "type": "object",
            "properties": {
                "Container": {
                    "type": "string"
                },
                "DurationNS": {
                    "type": "integer"
                },
                "FileSize": {
                    "type": "integer"
                },
                "Tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.gstreamerTrackDoc"
                    }
                }
            }
        },
...
        "api.gstreamerTrackDoc": {
            "type": "object",
            "properties": {
                "CapsName": {
                    "type": "string"
                },
                "Channels": {
                    "type": "integer"
                },
                "Codec": {
                    "type": "string"
                },
                "FrameRateDen": {
                    "type": "integer"
                },
                "FrameRateNum": {
                    "type": "integer"
                },
                "Height": {
                    "type": "integer"
                },
                "Index": {
                    "type": "integer"
                },
                "Language": {
                    "type": "string"
                },
                "PadName": {
                    "type": "string"
                },
                "Rate": {
                    "type": "integer"
                },
                "Title": {
                    "type": "string"
                },
                "Type": {
                    "type": "string"
                },
                "Width": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}